requirements.txt
requests==2.26.0
Django==3.2.5
djangorestframework==3.12.4
websockets==9.1
asgiref~=3.4.1
channels~=3.0.4
progressbar~=2.5
paho-mqtt~=1.5.1
toml~=0.10.2
pymongo==3.12.0
mongoengine~=0.23.1
pytz~=2021.1
django-cors-headers==3.8.0
channels-redis==3.3.0
opcua==0.98.13
confluent-kafka==1.7.0
zookeeper==1.3.2
python-dateutil~=2.8.2
openpyxl==3.0.9
pandas~=1.3.4
watchdog~=2.1.6
djangorestframework-simplejwt==5.0.0
django-filter==21.1
aiohttp==3.8.1
lxml==4.7.1